update to v1.0.0_RC1

+++ common/application/FlashHelper.h
@@ -1,14 +1,14 @@
 /**
  * @file FlashHelper.h
- * @brief Eve's connected flash helper functions
+ * @brief EVE's connected flash helper functions
  *
- * @author Tuan Nguyen <tuan.nguyen@brtchip.com>
+ * @author
  *
- * @date 2019
+ * @date 2024
  * 
  * MIT License
  *
- * Copyright (c) [2019] [Bridgetek Pte Ltd (BRTChip)]
+ * Copyright (c) [2024] [Bridgetek Pte Ltd (BRTChip)]
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -32,7 +32,7 @@
 #ifndef FLASH_HELPER__H
 #define FLASH_HELPER__H
 
-#include "EVE_Platform.h"
+#include "Common.h"
 
 typedef enum
 {
@@ -48,7 +48,7 @@ Flash_Cmd_Status_t FlashHelper_Update(EVE_HalContext *phost, uint32_t dest_flash
 Flash_Cmd_Status_t FlashHelper_Program(EVE_HalContext *phost, uint32_t dest_flash, uint32_t src_ram, uint32_t num);
 Flash_Cmd_Status_t FlashHelper_Read(EVE_HalContext *phost, uint32_t dest_ram, uint32_t src_flash, uint32_t num, uint8_t *read_data);
 void FlashHelper_Erase(EVE_HalContext *phost);
-uint8_t FlashHelper_GetState(EVE_HalContext *phost);
+uint32_t FlashHelper_GetState(EVE_HalContext *phost);
 Flash_Cmd_Status_t FlashHelper_Write(EVE_HalContext *phost, uint32_t dest_flash, uint32_t num, const uint8_t *write_data);
 uint32_t FlashHelper_SwitchState(EVE_HalContext *phost, uint8_t nextState);
 uint32_t FlashHelper_SwitchFullMode(EVE_HalContext *phost);
